+++ src/main/java/fr/maxlego08/essentials/commands/commands/items/CommandItemUnbreakable.java
@@ -14,7 +14,7 @@ public class CommandItemUnbreakable extends VCommand {
         super(plugin);
         this.setModule(fr.maxlego08.essentials.module.modules.ItemModule.class);
         this.setPermission(Permission.ESSENTIALS_ITEM_NAME);
-        this.setDescription(Message.DESCRIPTION_ITEM_NAME);
+        this.setDescription(Message.DESCRIPTION_ITEM_UNBREAKABLE);
         this.addBooleanOptionalArg("toggle");
         this.onlyPlayers();
     }